Move tests to unit test
